npd/release/src/requirements.txt
appdirs==1.4.0
numpy==1.12.0
packaging==16.8
pandas==0.19.2
progressbar2==3.12.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-utils==2.0.1
pytz==2016.10
six==1.10.0
xlrd==1.0.0
